backend/requirements.txt
# Core dependencies
Django==4.2.7                # Web framework
djangorestframework==3.14.0  # REST API framework
django-cors-headers==4.3.0   # Handle CORS
python-dotenv==1.0.0        # Environment variable management

# Database
psycopg2-binary==2.9.9      # PostgreSQL adapter
dj-database-url==2.0.0      # Database URL configuration

# Authentication
django-allauth==0.54.0      # Authentication
djangorestframework-simplejwt==5.3.0  # JWT authentication

# Task Queue
celery==5.3.1              # Asynchronous task queue
redis==4.6.0               # Redis for Celery backend

# API Documentation
drf-yasg==1.21.5           # Swagger/OpenAPI documentation

# Image Processing
Pillow==10.1.0             # For image handling

# Payment Processing
stripe==7.6.0              # Payment processing

# Production
gunicorn==21.2.0  # Production server
whitenoise==6.6.0  # Static file serving

# Storage
django-storages==1.14.2  # For handling file storage
boto3==1.29.3  # For AWS S3 storage

# Utilities
django-filter==23.3  # For filtering querysets
django-cleanup==8.0.0  # Automatically delete files
django-environ==0.11.2  # Environment variable handling
environs==10.3.0

# Email
sendgrid==6.10.0

# Development & Testing
setuptools<81            # drf-yasg still imports pkg_resources
pytest==7.4.0              # Testing framework
pytest-django==4.5.2       # Django test integration
black==23.7.0              # Code formatting
flake8==6.1.0             # Code linting
coverage==7.3.0           # Code coverage
